hecs

Struct Archetype

pub struct Archetype { /* private fields */ }

A collection of entities having the same component types

Accessing Archetypes is only required in niche cases. Typical use should go through the World.

Implementations

impl Archetype

pub fn has<T: Component>(&self) -> bool

Whether this archetype contains T components

pub fn has_dynamic(&self, id: TypeId) -> bool

Whether this archetype contains components with the type identified by id

pub fn get<'a, T: ComponentRef<'a>>(&'a self) -> Option<T::Column>

Borrow all components of a single type from these entities, if present

T must be a shared or unique reference to a component type.

Useful for efficient serialization.

pub fn len(&self) -> u32

Number of entities in this archetype

pub fn is_empty(&self) -> bool

Whether this archetype contains no entities

pub fn component_types(&self) -> impl ExactSizeIterator<Item = TypeId> + '_

Enumerate the types of the components of entities stored in this archetype.

Convenient for dispatching logic which needs to be performed on sets of type ids. For example, suppose you’re building a scripting system, and you want to integrate the scripting language with your ECS. This functionality allows you to iterate through all of the archetypes of the world with World::archetypes() and extract all possible combinations of component types which are currently stored in the World. From there, you can then create a mapping of archetypes to wrapper objects for your scripting language that provide functionality based off of the components of any given Entity, and bind them onto an Entity when passed into your scripting language by looking up the Entity’s archetype using EntityRef::component_types.

pub fn access<Q: Query>(&self) -> Option<Access>

How, if at all, Q will access entities in this archetype

pub fn satisfies<Q: Query>(&self) -> bool

Determine whether this archetype would satisfy the query Q

pub fn ids(&self) -> &[u32]

Raw IDs of the entities in this archetype

Convertible into Entitys with World::find_entity_from_id(). Useful for efficient serialization.