hecs

Struct EntityBuilder

pub struct EntityBuilder { /* private fields */ }

Helper for incrementally constructing a bundle of components with dynamic component types

Prefer reusing the same builder over creating new ones repeatedly.

let mut world = World::new();
let mut builder = EntityBuilder::new();
builder.add(123).add("abc");
let e = world.spawn(builder.build()); // builder can now be reused
assert_eq!(*world.get::<&i32>(e).unwrap(), 123);
assert_eq!(*world.get::<&&str>(e).unwrap(), "abc");

Implementations

impl EntityBuilder

pub fn new() -> Self

Create a builder representing an entity with no components

pub fn add<T: Component>(&mut self, component: T) -> &mut Self

Add component to the entity.

If the bundle already contains a component of type T, it will be dropped and replaced with the most recently added one.

pub fn add_bundle(&mut self, bundle: impl DynamicBundle) -> &mut Self

Add all components in bundle to the entity.

If the bundle contains any component which matches the type of a component already in the EntityBuilder, the newly added component from the bundle will replace the old component and the old component will be dropped.

pub fn build(&mut self) -> BuiltEntity<'_>

Construct a Bundle suitable for spawning

pub fn has<T: Component>(&self) -> bool

Checks to see if the component of type T exists

pub fn get<'a, T: ComponentRefShared<'a>>(&'a self) -> Option<T>

Borrow a shared reference T to some component type, if it exists

Takes a reference as its type parameter for consistency with EntityRef::get, even though it cannot be a unique reference because EntityBuilder does not perform dynamic borrow checking.

pub fn get_mut<'a, T: ComponentRef<'a>>(&'a mut self) -> Option<T>

Borrow a shared or unique reference T to some component type, if it exists

pub fn component_types(&self) -> impl Iterator<Item = TypeId> + '_

Enumerate the types of the entity builder’s components

pub fn clear(&mut self)

Drop previously added components

The builder is cleared implicitly when an entity is built, so this doesn’t usually need to be called.