hecs

Struct QueryBorrow

pub struct QueryBorrow<'w, Q: Query> { /* private fields */ }

A borrow of a World sufficient to execute the query Q

Note that borrows are not released until this object is dropped.

Implementations

impl<'w, Q: Query> QueryBorrow<'w, Q>

pub fn iter(&mut self) -> QueryIter<'_, Q>

Execute the query

pub fn view(&mut self) -> View<'_, Q>

Provide random access to the query results

pub fn iter_batched(&mut self, batch_size: u32) -> BatchedIter<'_, Q>

Like iter, but returns child iterators of at most batch_size elements

Useful for distributing work over a threadpool.

pub fn with<R: Query>(self) -> QueryBorrow<'w, With<Q, R>>

Transform the query into one that requires another query be satisfied

Convenient when the values of the components in the other query are not of interest.

Equivalent to using a query type wrapped in With.

Example

let mut world = World::new();
let a = world.spawn((123, true, "abc"));
let b = world.spawn((456, false));
let c = world.spawn((42, "def"));
let entities = world.query::<&i32>()
    .with::<&bool>()
    .iter()
    .map(|(e, &i)| (e, i)) // Copy out of the world
    .collect::<Vec<_>>();
assert_eq!(entities.len(), 2);
assert!(entities.contains(&(a, 123)));
assert!(entities.contains(&(b, 456)));

pub fn without<R: Query>(self) -> QueryBorrow<'w, Without<Q, R>>

Transform the query into one that skips entities satisfying another

Equivalent to using a query type wrapped in Without.

Example

let mut world = World::new();
let a = world.spawn((123, true, "abc"));
let b = world.spawn((456, false));
let c = world.spawn((42, "def"));
let entities = world.query::<&i32>()
    .without::<&bool>()
    .iter()
    .map(|(e, &i)| (e, i)) // Copy out of the world
    .collect::<Vec<_>>();
assert_eq!(entities, &[(c, 42)]);