hecs

Trait Bundle

pub unsafe trait Bundle: DynamicBundle { }

A statically typed collection of components

Bundles composed of exactly the same types are semantically equivalent, regardless of order. The interface of this trait is a private implementation detail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.