hecs

Trait DynamicBundle

pub unsafe trait DynamicBundle {
    // Provided method
    fn has<T: Component>(&self) -> bool { ... }
}

A dynamically typed collection of components

Bundles composed of exactly the same types are semantically equivalent, regardless of order. The interface of this trait, except has is a private implementation detail.

Provided Methods

fn has<T: Component>(&self) -> bool

Checks if the Bundle contains the given T:

let my_bundle = (0i32, 10.0f32);
assert!(my_bundle.has::<i32>());
assert!(my_bundle.has::<f32>());
assert!(!my_bundle.has::<usize>());

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.