hecs

Trait Query

pub trait Query {
    type Item<'a>;
}

A collection of component types to fetch from a World

The interface of this trait is a private implementation detail.

Required Associated Types

type Item<'a>

Type of results yielded by the query

This is usually the same type as the query itself, except with an appropriate lifetime.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.